src/main/java/com/github/brunomndantas/jscrapper/support/parser/single/text/primitive/SinglePrimitiveBooleanTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.single.text.primitive;

import com.github.brunomndantas.jscrapper.support.parser.single.text.SingleTextParser;
import org.openqa.selenium.WebElement;

public class SinglePrimitiveBooleanTextParser extends SingleTextParser {

    public SinglePrimitiveBooleanTextParser() {
        super(false);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Boolean.parseBoolean(element.getText());
    }

}